Use Web Crypto API for cookies and sessions
