+++ packages/react-router-node/globals.ts
@@ -6,6 +6,7 @@ import {
   Request as NodeRequest,
   Response as NodeResponse,
 } from "undici";
+import { webcrypto as nodeWebCrypto } from "node:crypto";
 
 declare global {
   namespace NodeJS {
@@ -24,6 +25,8 @@ declare global {
 
       ReadableStream: typeof ReadableStream;
       WritableStream: typeof WritableStream;
+
+      crypto: typeof nodeWebCrypto;
     }
   }
 
@@ -44,4 +47,6 @@ export function installGlobals() {
   global.fetch = nodeFetch;
   // @ts-expect-error - overriding globals
   global.FormData = NodeFormData;
+  // @ts-expect-error - overriding globals
+  global.crypto = nodeWebCrypto;
 }
