--- packages/react-router-node/implementations.ts
@@ -1,15 +1,0 @@
-import {
-  createCookieFactory,
-  createCookieSessionStorageFactory,
-  createMemorySessionStorageFactory,
-  createSessionStorageFactory,
-} from "react-router";
-
-import { sign, unsign } from "./crypto";
-
-export const createCookie = createCookieFactory({ sign, unsign });
-export const createCookieSessionStorage =
-  createCookieSessionStorageFactory(createCookie);
-export const createSessionStorage = createSessionStorageFactory(createCookie);
-export const createMemorySessionStorage =
-  createMemorySessionStorageFactory(createSessionStorage);
