+++ packages/react-router-node/package.json
@@ -35,13 +35,11 @@
   },
   "dependencies": {
     "@web3-storage/multipart-parser": "^1.0.0",
-    "cookie-signature": "^1.1.0",
     "source-map-support": "^0.5.21",
     "stream-slice": "^0.1.2",
     "undici": "^6.19.2"
   },
   "devDependencies": {
-    "@types/cookie-signature": "^1.0.3",
     "@types/source-map-support": "^0.5.4",
     "react-router": "workspace:*",
     "typescript": "^5.1.6"
