+++ packages/react-router/__tests__/server-runtime/cookies-test.ts
@@ -1,25 +1,4 @@
-import {
-  createCookieFactory,
-  isCookie,
-} from "../../lib/server-runtime/cookies";
-import type {
-  SignFunction,
-  UnsignFunction,
-} from "../../lib/server-runtime/crypto";
-
-const sign: SignFunction = async (value, secret) => {
-  return JSON.stringify({ value, secret });
-};
-const unsign: UnsignFunction = async (signed, secret) => {
-  try {
-    let unsigned = JSON.parse(signed);
-    if (unsigned.secret !== secret) return false;
-    return unsigned.value;
-  } catch (e: unknown) {
-    return false;
-  }
-};
-const createCookie = createCookieFactory({ sign, unsign });
+import { createCookie, isCookie } from "../../lib/server-runtime/cookies";
 
 function getCookieFromSetCookie(setCookie: string): string {
   return setCookie.split(/;\s*/)[0];
