+++ packages/react-router/__tests__/server-runtime/sessions-test.ts
@@ -1,38 +1,10 @@
-import { createCookieFactory } from "../../lib/server-runtime/cookies";
-import type {
-  SignFunction,
-  UnsignFunction,
-} from "../../lib/server-runtime/crypto";
-import {
-  createSession,
-  createSessionStorageFactory,
-  isSession,
-} from "../../lib/server-runtime/sessions";
-import { createCookieSessionStorageFactory } from "../../lib/server-runtime/sessions/cookieStorage";
-import { createMemorySessionStorageFactory } from "../../lib/server-runtime/sessions/memoryStorage";
+import { createSession, isSession } from "../../lib/server-runtime/sessions";
+import { createCookieSessionStorage } from "../../lib/server-runtime/sessions/cookieStorage";
+import { createMemorySessionStorage } from "../../lib/server-runtime/sessions/memoryStorage";
 
 function getCookieFromSetCookie(setCookie: string): string {
   return setCookie.split(/;\s*/)[0];
 }
-
-const sign: SignFunction = async (value, secret) => {
-  return JSON.stringify({ value, secret });
-};
-const unsign: UnsignFunction = async (signed, secret) => {
-  try {
-    let unsigned = JSON.parse(signed);
-    if (unsigned.secret !== secret) return false;
-    return unsigned.value;
-  } catch (e: unknown) {
-    return false;
-  }
-};
-const createCookie = createCookieFactory({ sign, unsign });
-const createCookieSessionStorage =
-  createCookieSessionStorageFactory(createCookie);
-const createSessionStorage = createSessionStorageFactory(createCookie);
-const createMemorySessionStorage =
-  createMemorySessionStorageFactory(createSessionStorage);
 
 describe("Session", () => {
   it("has an empty id by default", () => {
@@ -238,104 +210,6 @@ describe("Cookie session storage", () => {
       let setCookie2 = await storage.commitSession(session);
       expect(setCookie2).not.toEqual(setCookie);
     });
-  });
-});
-
-describe("Custom cookie-backed session storage", () => {
-  let memoryBacking = {};
-  let createCookieBackedSessionStorage =
-    createSessionStorageFactory(createCookie);
-  let implementation = {
-    createData(data) {
-      let id = Math.random().toString(36).substring(2, 10);
-      memoryBacking[id] = data;
-      return Promise.resolve(id);
-    },
-    readData(id) {
-      return Promise.resolve(memoryBacking[id] || null);
-    },
-    updateData(id, data) {
-      memoryBacking[id] = data;
-      return Promise.resolve();
-    },
-    deleteData(id) {
-      memoryBacking[id] = null;
-      return Promise.resolve(memoryBacking[id]);
-    },
-  };
-
-  it("persists session data across requests", async () => {
-    let { getSession, commitSession } = createCookieBackedSessionStorage({
-      ...implementation,
-      cookie: createCookie("test", { secrets: ["test"] }),
-    });
-    let session = await getSession();
-    session.set("user", "mjackson");
-    let setCookie = await commitSession(session);
-    session = await getSession(getCookieFromSetCookie(setCookie));
-
-    expect(session.get("user")).toEqual("mjackson");
-  });
-
-  it("returns an empty session for cookies that are not signed properly", async () => {
-    let { getSession, commitSession } = createCookieBackedSessionStorage({
-      ...implementation,
-      cookie: createCookie("test", { secrets: ["test"] }),
-    });
-    let session = await getSession();
-    session.set("user", "mjackson");
-
-    expect(session.get("user")).toEqual("mjackson");
-
-    let setCookie = await commitSession(session);
-    session = await getSession(
-      // Tamper with the session cookie...
-      getCookieFromSetCookie(setCookie).slice(0, -1)
-    );
-
-    expect(session.get("user")).toBeUndefined();
-  });
-
-  it('"makes the default path of cookies to be /', async () => {
-    let { getSession, commitSession } = createCookieBackedSessionStorage({
-      ...implementation,
-      cookie: createCookie("test", { secrets: ["test"] }),
-    });
-    let session = await getSession();
-    session.set("user", "mjackson");
-    let setCookie = await commitSession(session);
-    expect(setCookie).toContain("Path=/");
-  });
-
-  it("destroys sessions using a past date", async () => {
-    let spy = jest.spyOn(console, "warn").mockImplementation(() => {});
-    let { getSession, destroySession } = createCookieBackedSessionStorage({
-      ...implementation,
-      cookie: createCookie("test", { secrets: ["test"] }),
-    });
-    let session = await getSession();
-    let setCookie = await destroySession(session);
-    expect(setCookie).toMatchInlineSnapshot(
-      `"test=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT; SameSite=Lax"`
-    );
-    spy.mockRestore();
-  });
-
-  it("destroys sessions that leverage maxAge", async () => {
-    let spy = jest.spyOn(console, "warn").mockImplementation(() => {});
-    let { getSession, destroySession } = createCookieBackedSessionStorage({
-      ...implementation,
-      cookie: createCookie("test", {
-        maxAge: 60 * 60, // 1 hour
-        secrets: ["test"],
-      }),
-    });
-    let session = await getSession();
-    let setCookie = await destroySession(session);
-    expect(setCookie).toMatchInlineSnapshot(
-      `"test=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT; SameSite=Lax"`
-    );
-    spy.mockRestore();
   });
 });
